Handle stub-only modules
